Modified clientcert lookup
When a host is created through the Foreman web interface, it will set the
record's certname to nil, and this can cause a lookup failure when using a
Hiera backend (ex: Fancyass). So for the clientcert lookup only, we now check
to see if there is a host with a matching name as the passed clientcert value.
If there is, we check to see if its certname value is nil. If it is, we can
safely assume that this is the record we intended to find.